Skill menu for soldiers
This adds a skill menu for soldiers.
It will be displayed as an icon at the location of the special weapon icon in the battlescape.
Clicking it will open a moddable action menu. Each action can use an item or be completely scripted.
Skill actions can override weapon action costs. Item based actions can use items from the hand, the inventory
or a special weapon.